マイクロ秒単位で表示していたのをミリ秒単位に修正したのと、左右で異なるフォークを割り振っていた不具合を修正
